feat: sort by Overage in sales dashboard
