+++ api/organisations/models.py
@@ -400,13 +400,6 @@ class Subscription(LifecycleModelMixin, SoftDeleteExportableModel):
 
         add_single_seat(self.subscription_id)
 
-    def get_api_call_overage(self):
-        subscription_info = self.organisation.subscription_information_cache
-        overage = (
-            subscription_info.api_calls_30d - subscription_info.allowed_30d_api_calls
-        )
-        return overage if overage > 0 else 0
-
     def is_in_trial(self) -> bool:
         return self.subscription_id == TRIAL_SUBSCRIPTION_ID
